+++ NpgsqlRest/RoutineEndpoint.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Primitives;
+
 namespace NpgsqlRest;
 
 public struct RoutineEndpoint(
@@ -6,7 +7,7 @@ public struct RoutineEndpoint(
     Method method,
     RequestParamType requestParamType,
     bool requiresAuthorization,
-    string[] returnRecordNames,
+    string[] columnNames,
     string[] paramNames,
     int? commandTimeout,
     string? responseContentType,
@@ -16,15 +17,19 @@ public struct RoutineEndpoint(
     string? bodyParameterName,
     TextResponseNullHandling textResponseNullHandling,
     QueryStringNullHandling queryStringNullHandling,
-    HashSet<string>? authorizeRoles = null)
+    HashSet<string>? authorizeRoles = null,
+    bool login = false,
+    bool logout = false,
+    string? signInAuthenticationScheme = null)
 {
     internal HashSet<string> ParamsNameHash { get; } = new(paramNames);
     internal Action<ILogger, string, string, Exception?>? LogCallback { get; set; }
+
     public string Url { get; set; } = url;
     public Method Method { get; set; } = method;
     public RequestParamType RequestParamType { get; set; } = requestParamType;
     public bool RequiresAuthorization { get; set; } = requiresAuthorization;
-    public string[] ReturnRecordNames { get; set; } = returnRecordNames;
+    public string[] ColumnNames { get; set; } = columnNames;
     public string[] ParamNames { get; set; } = paramNames;
     public int? CommandTimeout { get; set; } = commandTimeout;
     public string? ResponseContentType { get; set; } = responseContentType;
@@ -35,4 +40,6 @@ public struct RoutineEndpoint(
     public TextResponseNullHandling TextResponseNullHandling { get; set; } = textResponseNullHandling;
     public QueryStringNullHandling QueryStringNullHandling { get; set; } = queryStringNullHandling;
     public HashSet<string>? AuthorizeRoles { get; set; } = authorizeRoles;
+    public bool Login { get; set; } = login;
+    public bool Logout { get; set; } = logout;
 }